Fold the untrusted-JSON rule into the conformance home [#1037]
The JSON parse-site rule shipped in its own file because ArchitectureConformanceTests
was gated by another change at the time. That left the repo with two plausible
homes for a structural rule, which is the drift the single-home doctrine exists
to prevent: the next person adding one now has to guess.

Everything moves into ArchitectureConformanceTests unchanged in what it asserts
- the rule, the three tables and their reasons, the gate-file assertion, the
vacuous-pass sentinel, the must-catch fixture with its adjacent must-not-catch
twin, and the code-reader rows. UntrustedJsonConformanceTests.cs is deleted; the
test count is unchanged on both frameworks, so no row was lost in the move.

Three things the move drops rather than carries, all of them duplicates of what
the destination already had: the file's own repo-root walk-up (the destination's
CallerFilePath helper resolves the root from a file one directory higher, and
the copy walked up a different number of levels for that reason), its own
IsBuildOutput, and its own code-line reader. The third is the one worth naming:
the copy dropped line comments, block openers, block continuations and XML-doc
lines, and so does the destination's OpensAComment, so HoldsAParseCall now calls
the shared SourceCallsInCode and the two readers cannot drift apart on the
question of what counts as a call.

One rename: TheScanRefusesAVacuousPass becomes TheJsonParseScanRefusesAVacuousPass.
In a file of one rule the old name was unambiguous; in a home with seventy it
names no scan in particular.

The two properties this move had to preserve were each made to fail on the
folded copy, then restored:

- the scan blinded to every file: the sentinel reports Assert.NotEmpty()
  Failure: Collection was empty, and the set-equality rule reddens beside it
- exact repo-relative paths replaced by bare file names: the set-equality rule
  reports Collections differ at index 0
- and the gate assertion, pointed at a production file that does not call the
  screen: "The gate file declared for SSO-Auth/Api/Oidc/DiscoveryJson.cs no
  longer calls the screen"

Gate on this commit: both target frameworks built with --warnaserror, 0
warnings; full suite 2697 passing on net9.0 and 2698 on net10.0, 0 failed, 4
skipped (the loopback-bind tests that need an elevation prompt, #1227).

Closes #1037
